feat: sign-up form API connection Fix API error message strange  path Change API service to have a main service (API) for all services uses
