+++ frontend/.gitignore
@@ -18,6 +18,9 @@
 # tailwind styles
 /public/styles/
 
+# languages
+/lang/compiled
+
 # misc
 .DS_Store
 *.pem
