+++ frontend/hoc/auth.ts
@@ -2,14 +2,14 @@ import { QueryClient } from 'react-query';
 
 import { dehydrate } from 'react-query/hydration';
 
-import USERS from 'services/users';
+import API from 'services/api';
 
 export function withProtection() {
   return async (context: any) => {
     const queryClient = new QueryClient();
     // remember to proxy cookies
     await queryClient.prefetchQuery('user', () =>
-      USERS.get('/', {
+      API.get('/', {
         headers: {
           Cookie: context.req.headers.cookie || '',
         },
