+++ frontend/lang/compiled/en.json
@@ -1,1 +1,2 @@
-{}
+{
+}
