+++ frontend/lang/compiled/es.json
@@ -1,1 +1,2 @@
-{}
+{
+}
