+++ frontend/lang/compiled/pt.json
@@ -1,1 +1,2 @@
-{}
+{
+}
