+++ frontend/layouts/protected-page/types.ts
@@ -1,0 +1,1 @@
+export type ProtectedProps = React.PropsWithChildren<React.ComponentProps<'div'> & {}>;
