+++ frontend/services/api.ts
@@ -8,35 +8,34 @@ const baseUrl = isServer
   ? process.env.NEXT_PUBLIC_BACKEND_HOST + process.env.NEXT_PUBLIC_API_URL
   : process.env.NEXT_PUBLIC_API_URL;
 
-const AUTHENTICATION = axios.create({
-  baseURL: `${baseUrl}/session`,
+const API = axios.create({
+  baseURL: baseUrl,
   headers: { 'Content-Type': 'application/json' },
-  xsrfCookieName: 'csrf_token',
-  xsrfHeaderName: 'X-CSRF-Token',
   withCredentials: true,
+  xsrfCookieName: 'csrf_token',
+  xsrfHeaderName: 'X-CSRF-TOKEN',
 });
 
 const onResponseSuccess = (response) => {
   try {
-    const parsedData = JSON.parse(response.data);
+    const parsedData = JSON.parse(response);
     return {
       data: dataFormatter.deserialize(parsedData),
       meta: parsedData.meta,
     };
   } catch (error) {
-    return response.data;
+    return response;
   }
 };
 
 const onResponseError = (error) => {
   if (error.response.data?.errors?.length > 0) {
-    return Promise.reject(new Error(error.response.data.errors));
+    return Promise.reject({ message: error.response.data.errors });
   }
 
-  // Do something with response error
   return Promise.reject(error);
 };
 
-AUTHENTICATION.interceptors.response.use(onResponseSuccess, onResponseError);
+API.interceptors.response.use(onResponseSuccess, onResponseError);
 
-export default AUTHENTICATION;
+export default API;
